Fixed create-wrms.sql to bring it up to current actual database.  Added a nofetch option to get-latest.sh which essentially just calls recreate-wrms.sh

+++ scripts/get-latest.sh
@@ -16,10 +16,14 @@ then
    exit 1
 fi
 
-echo "Grabbing latest version off $SITE"
-ssh $SITE "cd ~/wrms/scripts; ./dump-db.sh $DUMPDIR; cd $DUMPDIR; tar cvfz ~/wrms-tables.tgz t-*.sql"
-scp $SITE:wrms-tables.tgz ~/wrms/scripts/dump
-ssh $SITE "rm wrms-tables.tgz; rm -r $DUMPDIR"
+if [ "${SITE}" != "nofetch" ] ; then
+  echo "Grabbing latest version off $SITE"
+  ssh $SITE "cd ~/wrms/scripts; ./dump-db.sh $DUMPDIR; cd $DUMPDIR; tar cvfz ~/wrms-tables.tgz t-*.sql"
+  scp $SITE:wrms-tables.tgz ~/wrms/scripts/dump
+  ssh $SITE "rm wrms-tables.tgz; rm -r $DUMPDIR"
+else
+  echo "Using previous dump"
+fi
 pushd ~/wrms/scripts/dump
 tar xvfz wrms-tables.tgz
 pushd ~/wrms/scripts
